backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lighthouse-api"
version = "0.1.0"
description = "Lighthouse - Data Transformation Platform API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
    "duckdb>=1.1.0",
    "pandas>=2.2.0",
    "deltalake>=0.21.0",
    "google-cloud-storage>=2.18.0",
    "python-multipart>=0.0.12",
    "authlib>=1.3.0",
    "itsdangerous>=2.2.0",
    "httpx>=0.27.0",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "bandit>=1.7.0",
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "pip-audit>=2.7.0",
    "safety>=3.2.0",
    "coverage>=7.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "S", "B", "A", "C4", "UP", "RUF"]
ignore = ["S101", "S603", "S607"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/lighthouse_api"]
